Extract Parser class
